rohit_env3/mysite/requirements.txt
dj-database-url==0.5.0
Django==2.0.5
djangorestframework==3.8.2
gunicorn==19.8.1
psycopg2==2.7.4
pytz==2018.4
whitenoise==3.3.1
